dashboard/requirements.txt
# Django Framework
Django==5.0.1
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter (optional, for production)

# Authentication & Security
django-cors-headers==4.3.1

# API Communication
requests==2.31.0
httpx==0.25.2

# File Upload & Image Processing
Pillow==10.1.0

# Environment Variables
python-decouple==3.8

# Utilities
python-dateutil==2.8.2

# Development Tools
django-debug-toolbar==4.2.0

# Production Server (optional)
gunicorn==21.2.0

# Additional Django Extensions
django-extensions==3.2.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7
